Web+: don't pop DL window to the front
When you download a file, the Downloads window is opened to show
the progress etc. This is good, but once opened, it shouldn't
get activated (= pop to the front and getting focus) for each
new download. Often you've continued scrolling a webpage and when
the download finally starts, the DL window steals the focus...

Change-Id: Icdbd6bc5e2f0441eff650428834348f3fb508906
Reviewed-on: https://review.haiku-os.org/c/haiku/+/8799
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>
Reviewed-by: Jérôme Duval <jerome.duval@gmail.com>
Reviewed-by: Adrien Destugues <pulkomandy@pulkomandy.tk>
Haiku-Format: Haiku-format Bot <no-reply+haikuformatbot@haiku-os.org>