Background: Update when X or Y changed
Currently, in Background prefs, when you type a
value into the X or Y box, the preview is only
updated and the 'Apply' and 'Revert' buttons are
only enabled (or disabled if the value is changed
back) when you click into the other field. This
change updates once the user types a new value
into the boxes.

Change-Id: Ia3906d989639c16ec90d4ea157c4718061520bb0
Reviewed-on: https://review.haiku-os.org/c/haiku/+/8818
Reviewed-by: waddlesplash <waddlesplash@gmail.com>
Haiku-Format: Haiku-format Bot <no-reply+haikuformatbot@haiku-os.org>
Reviewed-by: Axel Dörfler <axeld@pinc-software.de>
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>