Reintroduce aligned-pointers optimization for strcmp.
It was initially added by tqh in ac827a2baa5e0c486b554cdee56c72624b499a40,
but was later removed in 39a81e5ac632d3356ebd4c0b40fb24aa3228833d
due to being rather broken.

This new version is loosely based on that design, but removes the
comparison logic out of the loop and lets the unaligned version
deal with the actual per-character comparison.

Saves around 0.1s out of 3.5s in the best case in a "jam -q HaikuDepot"
run with nothing to do (and that's only comparing with different
libroots on the same kernel; as this strcmp is used in the kernel as
well, we will benefit from this optimization there, too.)