HaikuControlLook: Fix ArrowShape color for Scrollbar buttons
Previously this would simply tint the passed in background color.
After hrev58595 this was changed to pass the proper Foreground color instead.

The codepath to disable/enable the color was missed in an oversight,
this puts back the original tint color for the foreground aswell.

Fixes: #19406

Change-Id: I6f75d24316e8ca3ecdb78f11fc8dede88f082780
Reviewed-on: https://review.haiku-os.org/c/haiku/+/8958
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>
Haiku-Format: Haiku-format Bot <no-reply+haikuformatbot@haiku-os.org>
Reviewed-by: waddlesplash <waddlesplash@gmail.com>