Interface kit: Be explicit about text color for DrawLabel
Previously NULL was the default argument for text labels,
this caused ControlLooks to pick "something" for the text color
In particular this previously caused mismatched color pairs.

Properly set this in the code to prevent this implicit
color choosing.

Api Clients can still pass = NULL directly if they
wish for the controlLook to think of a color itself.

Change-Id: Iea56e8d61269284f1d1470b852093bac5bf70de4
Reviewed-on: https://review.haiku-os.org/c/haiku/+/8963
Reviewed-by: John Scipione <jscipione@gmail.com>
Reviewed-by: nephele nephele <nep-git@packageloss.eu>
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>