XHCI: Adjust SMI disabling and only do it if the LEGSUP capability was found.
Otherwise we'll try to write to some random register at or past the end.

Adjust the bitmasks: rename XHCI_LEGCTLSTS_DISABLE_SMI to RESERVED_BITS
and define it in hex. Also add the "read-only" bits as a separate
declaration, and don't touch those either; then remove the setting
of the "events SMI" bits.