kernel/debug: Acquire the output mutex in the syslog sender and repeats checker.
This way we are less likely to encounter "Failed to acquire ... for
a long time", especially when onscreen debug output is enabled.

In the pending repeats checker, we wait at most 100ms to acquire the lock,
since this is run on the "kernel daemon" thread which we don't want
to block for too long.