kernel/vm: Forbid page faults in free_cached_pages().
Any that do occur in this state will be errors, but the page fault
handlers try to reserve pages to handle the fault even before checking
if the fault can be handled, so we could hit deadlocks or double-locks
before getting to the panic() without this.

Would have helped with diagnosing #12772.