kernel/vm: Don't release original cache references manually in cut_area.
There was one case in the old logic where we would double-release
references: if the secondCache->Resize() failed (or the Adopt(),
but this is much less likely to fail.)

Just reset the cacheChainLocker after we unlock all except the top cache.
It will unlock and release the reference as applicable.