kernel/vm: Resize the first new cache only after mapping the second succeeded.
Having a cache larger than its area is harmless, but resizing the
cache too early could lead to not having enough memory available
if we hit errors and need to resize it back later on, which would
be bad. So put the Resize() that shrinks the first area's cache
at the end, to avoid such a potential problem.