kernel/vm: Panic if we can't allocate a mapping object.
Failing here with ENOMEM means SIGSEGV sent to the process, which will
likely result in it getting killed.

All needed memory for mapping should be reserved upfront, or we should
at least fail earlier when trying to reserve pages for mapping. So
failure here indicates a problem with kernel memory accounting, and
shouldn't result in processes crashing, ideally.