ram_disk: Do not use cache_get_pages in Trim().
cache_get_pages locks and unlocks the cache to wait for busy pages,
as well as after it's done, which creates a race where some other
thread may start waiting on a page that we have marked busy. But
as we are about to free those pages, that wait will deadlock.

Just loop through the cache's pages directly, skipping any busy ones
and freeing the others directly.

Fixes a deadlock observed with BFS on a ramdisk.