kernel/x86_64: Handle LA57 paging in VMTranslationMap page table walks.
Without this, we will fail to free the last level of page tables.
This bug has apparently been present since LA57 was originally
introduced years ago, however it was likely not a problem before
now because only the BIOS bootloader has support for it, while
systems that support LA57 are more likely to be booted via EFI.

While at it, fix the address mask for non-LA57 in
DebugGetReverseMappingInfo.

Should fix #19460.