app_server: check glyph bounding box
AGG scanline storage is initialized with min and max coordinates with
maximum absolute values in the opposite direction, defining an invalid
rect that is updated when something is rendered. For a whitespace glyph
nothing is rendered, so we get an invalid bounding box. The moment we
offset that, the values overflow giving a valid enormous rect, which we
later use.

Fixes a crash in TextRenderer tests.

Change-Id: Ia8bbbcfd27bab6e6954eb890d9c783d04037794d
Reviewed-on: https://review.haiku-os.org/c/haiku/+/9129
Haiku-Format: Haiku-format Bot <no-reply+haikuformatbot@haiku-os.org>
Reviewed-by: waddlesplash <waddlesplash@gmail.com>