Interface Kit: Always use the specified "where" in ShowTip.
It was previously ignored if the tooltip was already open.

Note that, combined with the previous commit, this means that
the "where" passed in will basically always be used (except for
"sticky" tooltips.) Applications or classes that use this private
API but pass in wrong or bogus positions will now need to be fixed.