Interface Kit: Don't hide tooltip if we don't have any.
Ideally we would only hide our own tooltip and nobody else's, so
a TODO is added for that case.

This method is called for all mouse events if we get B_OUTSIDE_VIEW
or B_EXITED_VIEW as their "transit", but since tooltips themselves
get all mouse events and thus B_OUTSIDE_VIEW, this had the effect
of hiding all tooltips immediately instead of after the delay.