Tracker: Fix leak of addon menu items.
Every time the main menu was rebuilt, we leaked the "secondary"
add-on items, which included some FDs. Now, we just skip
creating addon items altogether if we're not going to insert them.

(BuildAddOnMenus() had logic to insert items, but this seems to
have been dead code, because "primary" would always be false,
so the primaryList would never have any items in it.)

Should fix #19506 and related problems.