libroot/posix: Switch back to using glibc's string/float-integer conversion routines.
This reverts commit 7174197a8da6549f719957d8152b5fced6aa7144
and all following changes to the related musl files.
(Besides just the reverts, though, it also uses glibc's
strtol-related routines even outside wchar_t.)

musl's behavior of only parsing POSIX-locale strings here may
be technically conformat, but in conjunction with a filled-in
"localeconv", produces behavior that glib does not expect,
and thus in locales with separators other than '.' and ',' as
in POSIX (or EN) locales, it doesn't work properly.

Fixes the main problem in #19582.