ICULocaleBackend: Validate that the created locale really exists.
Locale::createCanonical will happily accept all kinds of strings
that don't actually indicate valid locales, but down the line
ICU will behave strangely or even crash with these in certain methods.

So, use getISO3Language(), which will return nothing if the language
is unknown to ICU, to validate that the locale exists, or bail if
it doesn't.

Fixes #19576.