Installer: Add EFI Loader copy tool
Add a simple tool to the installer that copies the Haiku EFI loader
to a valid partition of the user's choice. Very basic with the
intention of making this process a bit less annoying; nothing more than a
one-to-one implementation of the copy instructions on this page:
https://www.haiku-os.org/guides/uefi_booting/

Also, fix bug in BDiskDeviceRoster::FindPartitionByMountPoint that caused
an error to be incorrectly returned on success.

Change-Id: Ifd3bb96d907dc93908de5ead007c3d5e72b110b8