IPv4 & TCP: Implement naive Path MTU Discovery.
 * Set IP_DONTFRAG on all TCP sockets by default.

 * Handle receiving B_NET_ERROR_MESSAGE_SIZE, update MaxSegmentSize
   appropriately, and trigger retransmit.

Tested and verified as working.

Part of #1073. However this doesn't implement all of what's needed,
only the most basic form of the algorithm (we don't record MTUs
for reuse later for instance, right now we only store routing
information for the LAN, it appears, so we will need somewhere else
to store that.)

Change-Id: I37a24b50db18af908c6f6257c6fb6d72127d2787
Reviewed-on: https://review.haiku-os.org/c/haiku/+/9402
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>
Reviewed-by: Jérôme Duval <jerome.duval@gmail.com>
Reviewed-by: waddlesplash <waddlesplash@gmail.com>