kernel/fs: Remove special handling for O_DISCONNECTED in remove_fd().
This method is used in close_fd_index(), the normal path for close()
operations to go through. So, if a close() had been called on a
disconnected FD, before this commit we would just leak it.

This change means that close_fd() and put_fd() are called on
disconnected FDs being close()d. That's the same set of operations
that deleting an I/O context does, though, so this should
hopefully not cause any problems.