Tracker: Draw dragged pose text selected, fixes #18572
Note this is a change to how BeOS R5 and Haiku has always worked.

Drawing dragged poses as selected ensures that the text can be read
even on black on black or white on white because the selection rect
creates contrast.

Lessen contrast by blending with gray when selected and cut, even
more when window is unfocused.

Pose text on Desktop is inverted compared to in a folder by default
so this bug typically occurs when dragging to and from Desktop.

Change-Id: Icb1ee7ccbf9b5e2d8c3a4b89c6b90e19da15c917
Reviewed-on: https://review.haiku-os.org/c/haiku/+/9520
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>
Reviewed-by: humdinger humdinger <humdinger@mailbox.org>
Reviewed-by: Adrien Destugues <pulkomandy@pulkomandy.tk>