kernel/block_cache: Add fast-paths for getting and putting blocks.
Using the new rw_lock, plus a spinlock for unused blocks.

This massively reduces lock contention on "git status". With
a cold disk cache (just after reboot) in a 4-core VM, "git status"
in the buildtools directory goes from ~33.4s down to ~20.1s, and with
a hot disk cache, from ~13.4s -> ~9.1s. Total time spent waiting on
the main lock goes from ~9-10 minutes down to ~5 in the cold case,
and in the hot case, down to < 1 second (instead there's ~2min
of wait time on the entry cache lock.)

It is probably worth noting that the unused spinlock is critical
to this working out. Without it, we still hit the fast path
hundreds of thousands of times in "git status", but the unlock-relock
happens often enough that it makes performance worse: I saw times of
40-50 seconds instead of 30 with that setup.

Change-Id: Ica8223ab6f1aacf65aa9b7408d980bb85c758ab9
Reviewed-on: https://review.haiku-os.org/c/haiku/+/9531
Reviewed-by: waddlesplash <waddlesplash@gmail.com>
Reviewed-by: Axel Dörfler <axeld@pinc-software.de>